Split up tests, gofmt, goimports, docs
