+++ shared/bls/bls12/docs.go
@@ -1,4 +1,6 @@
-// Package bls implements a go-wrapper around a library implementing the
+// Package bls12 implements a go-wrapper around a library implementing the
 // the BLS12-381 curve and signature scheme. This package exposes a public API for
 // verifying and aggregating BLS signatures used by Ethereum 2.0.
+//
+// This implementation uses the library written by Herumi.
 package bls12
